+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 -   **terra-syntax-editor** has been removed. Use [ck-editor](https://ckeditor.com/ckeditor-4/) instead.
 -   **dependencies** removed peer dependency `ng2-ace-editor` since we no longer depend on it.
+-   **terra-tag** removed deprecated `onCloseTag` output.
+-   **terra-taglist** removed deprecated `onCloseTag` output.
 -   **terra-base-toolbar**'s input `inputIsBreadcrumbs` has been removed. There's no replacement.
 -   **file-input** removed attribute `_overlay` and method `showFileBrowser()`.
+-   **terra-label-tooltip** directive has been removed. Use `TooltipDirective` with a set `onlyEllipsisTooltip` input instead.
 
 # 5.10.1 (Unreleased)